Move get_login_session_form to user_authn.
